[experiment] opt-in clang-cl support for Windows
Adds an experimental, opt-in build path for CoreCLR's JIT (clrjit.dll)
on Windows using LLVM clang-cl instead of MSVC cl.exe.

Usage:

    # Direct (forces Ninja generator):
    src\coreclr\build-runtime.cmd -clangcl -release -component jit

    # Via top-level build (set env var, then build normally):
    $env:__UseClangCl="1"
    build.cmd Clr.Jit -c Release /p:NoPgoOptimize=true

Requires clang-cl.exe on PATH or under %ProgramFiles%\LLVM\bin
(winget install LLVM.LLVM).

Build infrastructure:
* src/coreclr/build-runtime.cmd: new -clangcl flag, also honors a
  pre-set __UseClangCl=1 env var (so MSBuild can opt in without
  modifying the call into build-runtime.cmd).
* eng/native/gen-buildsys.cmd: when __UseClangCl=1, locates
  clang-cl.exe and injects -DCMAKE_C_COMPILER /
  -DCMAKE_CXX_COMPILER. Forces the Ninja generator.
* eng/native/configurecompiler.cmake: a 10-line clang-cl branch
  inside the existing if (MSVC) block. Each suppression was
  empirically verified to actually fire by replacing -Wno-foo with
  -Wno-error=foo and rebuilding (see comment block).

Source fixes (all MSVC-safe; verified MSVC build still passes):
* src/coreclr/utilcode/pedecoder.cpp: remove a stray line-continuation
  '\' that silently merged two STATIC_CONTRACT_* statements (this
  was actually masking a clang frontend crash). Also add explicit
  cast for __unaligned char[] -> const char*.
* src/coreclr/jit/gentree.cpp: add missing 'id =' on a compare-
  greater-than scalar branch where the ?: result was being discarded
  (a real JIT bug found by -Wunused-value).
* src/coreclr/jit/importercalls.cpp: replace two 'impPopStack().val;'
  pop-and-discard statements with one 'impPopStack(2);'.
* src/coreclr/jit/utils.cpp: explicit (int) cast on size_t template
  argument compared against int (sign-compare).
* src/coreclr/inc/sigparser.h: 'operator= = default' to satisfy
  -Wdeprecated-copy-with-user-provided-copy.
* src/coreclr/utilcode/debug.cpp: __assume(0) after
  RaiseFailFastException so clang sees FailFastOnAssert as truly
  noreturn (the WinAPI declaration isn't __declspec(noreturn)).
* src/coreclr/utilcode/fstring.cpp: gate '#pragma optimize("t", on)'
  on _MSC_VER && !__clang__ (clang-cl only accepts an empty arg).
* src/coreclr/utilcode/stacktrace.cpp: reinterpret_cast<LPVOID>(...)
  for FARPROC -> LPVOID conversion plus (int) cast on ARRAY_SIZE
  for the loop counter.
* src/native/minipal/guid.c: (unsigned int) cast on guid.Data1 for
  %08x format spec.
* src/native/minipal/xoshiro128pp.c: 'size_t i' instead of 'int i'
  for the sizeof-based loop bound.

Verified:
* MSVC build.cmd Clr.Jit -c Release /p:NoPgoOptimize=true passes
* clang-cl __UseClangCl=1 build.cmd Clr.Jit -c Release
  /p:NoPgoOptimize=true passes (clang 22.1.5)
* JIT smoke test (loop sum, Vector<int> SIMD, throw/catch) JITs and
  runs correctly with the clang-cl-built clrjit.dll, exit code 100,
  identical output to the MSVC-built JIT.

Out of scope for this PR (intended as follow-ups):
* Other CoreCLR components (coreclr.dll, DAC, mscordbi, host).
* Libs native and host clang-cl support.
* VS-integrated -T ClangCL MSBuild path.
* CI pipeline coverage.

+++ src/coreclr/jit/utils.cpp
@@ -4184,7 +4184,7 @@ public:
             return;
         }
 
-        if (strBufferSize > bufferSize)
+        if (strBufferSize > (int)bufferSize)
         {
             m_pBuffer = new WCHAR[strBufferSize];
         }
