+++ src/native/minipal/xoshiro128pp.c
@@ -29,7 +29,7 @@ static void jump(struct minipal_xoshiro128pp* pState) {
 	uint32_t s1 = 0;
 	uint32_t s2 = 0;
 	uint32_t s3 = 0;
-	for (int i = 0; i < sizeof JUMP / sizeof * JUMP; i++)
+	for (size_t i = 0; i < sizeof JUMP / sizeof * JUMP; i++)
 		for (int b = 0; b < 32; b++) {
 			if (JUMP[i] & UINT32_C(1) << b) {
 				s0 ^= s[0];
